chore(workspace): drop openinfer-vllm-support, inline tokenizer loader into consumers
The crate was an 81-line sync bridge over vllm-text's async model-file
resolution (global OnceCell<Mutex<Runtime>> + nested-runtime guard +
TokenizerSource dispatch). Its three real consumers now each own the same
~30-line loader, with test call sites untouched:

- openinfer-qwen3 tests/common (9 test files)
- openinfer-qwen35 tests/common (3 test files)
- bench_serving binary via a new bench_serving/tokenizer.rs module

Also removes dead weight uncovered while auditing consumers:

- openinfer-server/tests/common/mod.rs: orphan file, no test target or
  `mod common;` declaration anywhere, never compiled
- workspace once_cell: openinfer-vllm-support was the only user
- CI unit-test matrix + clippy entries for the deleted crate

Changes vs the deleted crate: the global runtime singleton becomes a
throwaway per-call current-thread runtime (tokenizer is loaded once per
process; the singleton was premature), and qwen3 dev-tokio gains
net/time to preserve remote model-id resolution parity.
