--- openinfer-vllm-support/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-edition = "2024"
-license = "Apache-2.0"
-name = "openinfer-vllm-support"
-version = "0.1.0"
-
-[dependencies]
-once_cell = { workspace = true }
-tokio = { workspace = true, features = ["net", "rt", "time"] }
-vllm-text = { workspace = true }
-vllm-tokenizer = { workspace = true }
-
-[lints]
-workspace = true
